Remove wrong casts a la previous commit
There were no new runtime bugs to find here, but they could conceal bugs
in the future. Some type checkers will widen literals when assigning in
an init, so we clarify we keep the self.justify/overflow as the string
literals in order to keep type checking as strict as possible
everywhere.