replace `# type: ignore` w/ `TYPE_CHECKING` guard
While `type: ignore` indicates errors on the line should be ignored, it
doesn't stop all type checkers from inferring the type of `_attr_module`
could possibly be None. Instead of the ignore, hide the assignment
behind a TYPE_CHECKING guard.

This type of guard really is needed - we cannot simply use an assert
within a TYPE_CHECKING guard because in the future when annotations are
deferred stringized (forward references) will become ambiguous. See:
https://github.com/microsoft/pyright/issues/2442