declare type of `renderables`
Because renderables is later assigned to a list of `Styled` objects, a
type checker won't necessarily know that you don't intend the list to
become the more narrow `List[Styled]` type or the wider
`List[ConsoleRenderable]`; if the former the code later adding a
`ConsoleRenderable` would be wrong. We declare the type of renderables
to signal to type checkers its proper type.