replaced always-false evaluation
It's not clear to static analysis tools `value` is always assigned here,
but when looking at it I realized this is always False so it can be
replaced with False anyway. I've adjusted the docstring accordingly.

This works because `bool(value)` is evaluated on the last value of the
iterable if and only if the iterable consists entirely of None;
therefore, it is False.