Signal type checker obj can't be type `type` here
This is a complicated one. The only reason mypy doesn't complain here is
because of limitations in its inference ability; if it was better it
would correctly complain.

The problem is we check `elif isinstance(obj, _CONTAINERS)` where
`_CONTAINERS` is a tuple of types; this cannot help us distinguish if
`obj` is a type `type` or not. Why does this matter? We unintentionally
statically declared it could be type `type` on a previous `elif` line of
the sort `if A and not isinstance(obj, type)`. As a result, in
subsequent lines we have indicated to the type checker the logical
negation of this: `obj` is of type `Any | type`. But the `isinstance`
check - by coincidence of the way it was defined in a `dict` with keys
of type `type` - cannot narrow away the possibility that `obj` is of
type `type`, therefore we need to assert this to be the case.