Fix _lru_cache typing
This class had a number of different errors in pyright that arguably
only pass in mypy due to limitations in it. When using OrderedDict's
classes without `super` the type checker can't necessarily infer the
classes should be compatible with it, and the value setting line in
`__getitem__` was an example where you will get an error in pyright
because the subclass relationship can't be unambiguously inferred.
Although it was only strictly necessary in that location, I've replaced
the use of OrderedDict with super() throughout the class to be
consistent and properly support the typing needed for pyright and any
other stricter type checkers in the future.

Several hints are needed for `self` and the type of value. None of these
will be needed when rich drops support for 3.6 as we will be able to use
the OrderedDict in the typing module.