Fix and narrow repr.py typing
First of many stabs at eliminating Any and # type: ignore from the
codebase. Here we eliminated several different type ignores; the
remaining # type: ignore is just a mypy limitation.

I also narrowed the typing of RichReprResult from Iterable[Any|...]] to
Iterable[object] since after inspection of the code this is what's
really required.

There's one particularly nasty cast that's needed. It's definitely a
mypy bug that it's not throwing an error for us here. See:
https://github.com/python/mypy/issues/1178 and I've opened a similar
issue for pyright. There's an open PR for mypy that would resolve this,
but at least until this is fixed we need to be casting.

In order to correctly type the auto_repr function, I introduced a
SupportsRichRepr protocol; we could also have made this runtimecheckable
and replaced a hasattr check in the code with it, but it seemed
unnecessary so I opted not to.