Add support for a "fuzzy" choice matcher
When choices are pretty distinct, it is nice to let the user only type
part of the choice and let us automatically match the rest. For example
if the choices are "Strike" and "Ball", then if the user types "b" we
know they meant ball.

This commit adds a new Prompt subclass to support this fuzzy matching.