Experimenting with the __rich_console__ protocol
Just a little experiment to see if I can feel my way around the innards of
Rich and, in this case, get a feel for doing fun things with the
__rich_console__ protocol.

The code here isn't tidy yet, this is just a quick experimental hack to see
if I can wrap my head around things. It lacks docstrings and the like for
starters. It's also missing a type hint for the case_filter base constructor
parameter, mostly because, last thing at night, I had a crisis of confidence
that I can recall how to Callable type an instance method as opposed to a
normal function.

Like... is str.upper Callable[[],str] or something else? Kinda feels like it
might be Callable[[str],str] but that would be what:

def foo( thing: str ) -> str:
   ...

would be and... yeah, rabbit hole at gone 23:00 on a school night. I'll
ponder it tomorrow.