Fix precision on sizes smaller than base in rich.filesize
In the base set in decimal(), this change will only affect the bytes file size. The example in the doc-string has been edited to showcase this precision as well. I also removed the prevision showcase from Example 2, as I felt it best to separate the examples to one change only. I.e., no args, separator arg, and precision arg.

Note that this change with the suffixes list means it may affect user code. The size == 1 code was removed to not assume size == 1 under the specified base means bytes. The code as it stands now will more implicitly return `"1.0 bytes"` when size == 1 (assuming the precision of 1).

You may prefer an alternative suffix to `bytes`, i.e., `b` or `B`, but I chose `bytes` since it's what was already used in the more explicit returns that are now removed.

Closes #2951