Swap open-vs-close tag order when converting Text to markup.
I'm torn on whether this is a pleasingly simple solution or a solution
that's too simple to possibly be correct. Still, it makes the test pass,
so it'll do as a first cut while I decide if I want to make a second,
more complex run at it.

(Possibly track the order open-tags were rendered to ensure that
multiple close-tags at a given offset get popped off in reverse open-tag
order?)