Enable per-column enabling or disabling of highlighting
Defaults to the value of the self.table when the default value of None is unchanged.

It seems like the Column constructor didn't receive a `highlight` value, so I guess that it used to receive a highlight value at a later point. Someone who knows the project better could comment on that.