Optimise '_get_codepoint_cell_size'.
This optimisation passes rich testing (obviously) and it also passes the full Textual test suite.

Replace explicit binary search with the module 'bisect'. To simplify the code, we replace the list of ranges and widths present in rich/_cell_widths.py with the lists of range starts, range ends, and respective widths.
Having a separate list for the starts makes it easier to apply the bisection.