Optimise 'set_cell_size'.
This optimisation passes Textual testing as well.

Instead of repeatedly computing the cell length of the current attempted split, we compute the width of each character and use binary search to find the cut point.
This is faster because we only compute the length of the full string once (as we compute the width of each character once) instead of recomputing cell lengths.