when using inline HTML styles, directly apply rules to anchors
When exporting captured output to HTML, using inline CSS styles,
linked text is formatted as:
       <span style="..."><a href="...">text</a></span>

It seems that Web browsers will then render the linked text
with the default style for unvisited links (typically blue).

The CSS rules are, however, correctly applied if put
directly on the anchor element:
       <a style="..." href="...">text</a>

Tested with Firefox 128.4 and Google Chrome 131.