fix(traceback): highlight win32 paths
`PathHighlighter` uses a regex to parse paths for highlighting.
It used `/` as the final path component delimiter. This diff
changes that to "either `/` or `\`".

I toyed with using `os.path.sep`, but there's no reason to *not*
highlight paths from other OSes than the one we're running on
at the moment.